refactor(renderer): extract bootstrap into its own module
Move the script-loading concern (loadVizGlobal + dynamic plantuml.js import)
into bootstrap.ts so renderer.ts contains only the rendering pipeline.
Tests now vi.mock("./bootstrap") to substitute the loader, which removes
the test-only setPlantUMLModuleForTests export from production code.
